RS-0: Fix rspec tests
