+++ .rspec
@@ -1,1 +1,1 @@
---require spec_helper
+--require rails_helper
