+++ spec/models/role_spec.rb
@@ -1,5 +1,3 @@
-require 'rails_helper'
-
 RSpec.describe Role, type: :model do
   pending "add some examples to (or delete) #{__FILE__}"
 end
