+++ spec/models/user_spec.rb
@@ -1,5 +1,3 @@
-require 'rails_helper'
-
 RSpec.describe User, type: :model do
   pending "add some examples to (or delete) #{__FILE__}"
 end
